refactor: add fare-contracts as a module
