--- src/fare-contracts/index.ts
@@ -1,4 +1,0 @@
-export {FareContractAndReservationsList} from './FareContractAndReservationsList';
-export {DetailsContent} from './details/DetailsContent';
-export {getValidityStatus} from './utils';
-export type {UserProfileWithCount} from './types';
