+++ src/modules/fare-contracts/components/ActivateNowSectionItem.tsx
