+++ src/modules/fare-contracts/components/FareContractDetailItem.tsx
