+++ src/modules/fare-contracts/components/FareContractFromTo.tsx
@@ -1,8 +1,8 @@
 import {ContrastColor} from '@atb-as/theme';
 import {type RecentFareContractType} from '@atb/recent-fare-contracts';
 import {FareContractType, TravelRightDirection} from '@atb-as/utils';
-import {ZonesFromTo} from '@atb/fare-contracts/components/ZonesFromTo';
-import {HarborsFromTo} from '@atb/fare-contracts/components/HarborsFromTo';
+import {ZonesFromTo} from './ZonesFromTo';
+import {HarborsFromTo} from './HarborsFromTo';
 import {
   findReferenceDataById,
   useFirestoreConfigurationContext,
