+++ src/modules/fare-contracts/components/HarborsFromTo.tsx
@@ -1,6 +1,6 @@
 import {TravelRightDirection} from '@atb-as/utils';
 import {ContrastColor} from '@atb-as/theme';
-import {BorderedFromToBox} from '@atb/fare-contracts/components/BorderedFromToBox';
+import {BorderedFromToBox} from './BorderedFromToBox';
 import {useHarbors} from '@atb/harbors';
 
 type HarborsProps = {
