+++ src/modules/fare-contracts/components/InspectionSymbol.tsx
