+++ src/modules/fare-contracts/details/DetailsContent.tsx
@@ -12,9 +12,9 @@ import {
   getFareContractInfo,
   hasShmoBookingId,
   mapToUserProfilesWithCount,
-} from '@atb/fare-contracts/utils';
+} from '../utils';
 import {useMobileTokenContext} from '@atb/mobile-token';
-import {OrderDetailsSectionItem} from '@atb/fare-contracts/sections/OrderDetailsSectionItem';
+import {OrderDetailsSectionItem} from '../sections/OrderDetailsSectionItem';
 import {
   GenericSectionItem,
   LinkSectionItem,
@@ -41,7 +41,7 @@ import {ConsumeCarnetSectionItem} from '../components/ConsumeCarnetSectionItem';
 import {ActivateNowSectionItem} from '../components/ActivateNowSectionItem';
 import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 import {formatPhoneNumber} from '@atb/utils/phone-number-utils';
-import {UsedAccessesSectionItem} from '@atb/fare-contracts/details/UsedAccessesSectionItem';
+import {UsedAccessesSectionItem} from './UsedAccessesSectionItem';
 import {ShmoTripDetailsSectionItem} from '@atb/mobility/components/ShmoTripDetailsSectionItem';
 import {FareContractHeaderSectionItem} from '../sections/FareContractHeaderSectionItem';
 import {FareContractShmoHeaderSectionItem} from '../sections/FareContractShmoHeaderSectionItem';
