+++ src/modules/fare-contracts/sections/FareContractHeaderSectionItem.tsx
@@ -7,15 +7,15 @@ import {StyleSheet, useThemeContext} from '@atb/theme';
 import {useGetPhoneByAccountIdQuery} from '@atb/on-behalf-of/queries/use-get-phone-by-account-id-query';
 import {useAuthContext} from '@atb/auth';
 import {formatPhoneNumber} from '@atb/utils/phone-number-utils';
-import {FareContractFromTo} from '@atb/fare-contracts/components/FareContractFromTo';
-import {Description} from '@atb/fare-contracts/components/FareContractDescription';
-import {ValidTo} from '@atb/fare-contracts/components/ValidTo';
+import {FareContractFromTo} from '../components/FareContractFromTo';
+import {Description} from '../components/FareContractDescription';
+import {ValidTo} from '../components/ValidTo';
 import {useFetchOnBehalfOfAccountsQuery} from '@atb/on-behalf-of/queries/use-fetch-on-behalf-of-accounts-query';
 import {MessageInfoBox} from '@atb/components/message-info-box';
-import {WithValidityLine} from '@atb/fare-contracts/components/WithValidityLine';
-import {ProductName} from '@atb/fare-contracts/components/ProductName';
-import {ValidityTime} from '@atb/fare-contracts/components/ValidityTime';
 import {isSentOrReceivedFareContract} from '@atb/ticketing';
+import {WithValidityLine} from '../components/WithValidityLine';
+import {ProductName} from '../components/ProductName';
+import {ValidityTime} from '../components/ValidityTime';
 
 type Props = {
   fareContract: FareContractType;
