+++ src/modules/fare-contracts/use-validity-line-colors.ts
