+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_FareContractsScreen.tsx
@@ -1,8 +1,8 @@
 import {FullScreenHeader} from '@atb/components/screen-header';
 import {ThemeText} from '@atb/components/text';
 import {useFirestoreConfigurationContext} from '@atb/configuration';
-import {DetailsContent} from '@atb/fare-contracts';
-import {FareContractOrReservation} from '@atb/fare-contracts/FareContractOrReservation';
+import {DetailsContent} from '@atb/modules/fare-contracts';
+import {FareContractOrReservation} from '@atb/modules/fare-contracts';
 import {findReferenceDataById} from '@atb/configuration';
 import {StyleSheet, Theme} from '@atb/theme';
 import {Reservation} from '@atb/ticketing';
